Fork `dominate` into `domin8`:
I have been using `dominate` for a few of my projects and have been wanting
more features from it, but development seems to have slowed down, and there are
lots of open issues and issues that were closed which I'd like to see
implemented.

My main goals for this fork are the following:
* Drop support for EOL Python versions 3.7 and below
* Remove legacy Python 2 compatibility
* Add better typing support
* Add support for HTMX attributes as keyword args
* Improve testing and linting setup

I intend to keep the code style and structure mostly the same so that these
changes can be upstreamed in the future.

Some side goals include compiling `domin8` to Mypyc, though I don't know if
this will bring a noticable performance impact, if any.