Add support for HTMX attributes:
Previously only `data_` and `aria_` kwargs where converted to the common dash
delimited versions (like `data-xyz`), meaning that other attributes like `hx-`
used by HTMX would not work, and would require you to set the attributes on
the tag using `tag["hx-abc"] = "xyz"`. This commit allows you to use `hx_abc`
(for example) as a keyword argument, meaning you can now use HTMX attributes
inline.

This is a breaking change for anyone expecting that `some_attr` be literally
rendered as `some_attr`, but the common style for HTML attributes is to use
dashes and not underscores (like how the `data` and `aria` tags work).

There should probably be a way to additionally specify prefixes to ignore, but
this should suffice for now.