Add `_` attribute special case to support Hyperscript:
This is a follow up to last commit, but it turns out using `_` as a kwargs was
already broken since dominate would strip leading underscores in order to
support attributes like `_class`. This meant that, when rendered, using just
`_` as an attribute would result in `="whatever"` being rendered, notice the
absence of an attribute name.