Fix `domin8` not being able to be used as a package name:
For whatever reason a single `__init__.py` file is not enough to import all
the dominate package info, I have to emulate the package structure and export
the corresponding fields for each one. This isn't ideal, but I don't know how
else to do it. There might be a way to do this directly in setuptools, but I
cannot seem to find out how.